+++ ft_memset.c
@@ -6,20 +6,20 @@
 /*   By: gateixei <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/06 17:46:58 by gateixei          #+#    #+#             */
-/*   Updated: 2021/12/10 17:16:10 by gateixei         ###   ########.fr       */
+/*   Updated: 2021/12/14 19:28:29 by gateixei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memset(void *b, int c, int len)
+void	*ft_memset(void *b, int c, size_t len)
 {
 	unsigned char	*str;
 	int				i;
 
 	i = 0;
 	str = (unsigned char *) b;
-	while (i < len)
+	while ((size_t) i < len)
 		str[i++] = c;
 	return ((void *) &str[0]);
 }
